fix: decode remainingPercentage as Double to support Minimax quota display
Minimax returns remainingPercentage as a floating-point value (e.g. 93.46),
causing JSONDecoder to fail when the field was typed as Int?, which silently
dropped the entire quota entry. Changed to Double? and cast to Int at display.

Bump version 0.1.4 → 0.1.5

+++ 9QuoterApp/Sources/9QuoterApp/Models/Models.swift
@@ -30,7 +30,7 @@ struct QuotaEntry: Codable {
     let remaining: Int
     let unlimited: Bool
     let resetAt: String?
-    let remainingPercentage: Int?
+    let remainingPercentage: Double?
 }
 
 struct ExtraUsage: Codable {
